+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
  - Default web conferencing path to processor address
+ - Volume button goes red when volume is muted
 
 ## [3.2.0] - 2018-05-09
 ### Changed